fix: Correct weighted_score type in RPC function
Changed weighted_score from INTEGER to NUMERIC(10,2) to match
the actual column type in tournament_clips table.

+++ supabase/sql/optimize-for-scale.sql
@@ -123,7 +123,7 @@ RETURNS TABLE (
   genre TEXT,
   slot_position INTEGER,
   vote_count INTEGER,
-  weighted_score INTEGER,
+  weighted_score NUMERIC(10,2),
   view_count INTEGER,
   created_at TIMESTAMPTZ,
   has_voted BOOLEAN,
